feat: Refactor dictionary service.
