--- src/main/java/org/spin/base/util/DictionaryUtil.java
@@ -1,107 +1,0 @@
-/*************************************************************************************
- * Product: Adempiere ERP & CRM Smart Business Solution                              *
- * This program is free software; you can redistribute it and/or modify it           *
- * under the terms version 2 or later of the GNU General Public License as published *
- * by the Free Software Foundation. This program is distributed in the hope          *
- * that it will be useful, but WITHOUT ANY WARRANTY; without even the implied        *
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.                  *
- * See the GNU General Public License for more details.                              *
- * You should have received a copy of the GNU General Public License along           *
- * with this program; if not, write to the Free Software Foundation, Inc.,           *
- * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA.                            *
- * For the text or an alternative of this public license, you may reach us           *
- * Copyright (C) 2012-2018 E.R.P. Consultores y Asociados, S.A. All Rights Reserved. *
- * Contributor(s): Yamel Senih www.erpya.com                                         *
- *************************************************************************************/
-package org.spin.base.util;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import org.compiere.util.Util;
-
-/**
- * Class for handle records utils values
- * @author Yamel Senih, ysenih@erpya.com , http://www.erpya.com
- */
-public class DictionaryUtil {
-	
-	public static String ID_PREFIX = "_ID";
-
-	public static String TRANSLATION_SUFFIX = "_Trl";
-
-	/**
-	 * Get Context column names from context
-	 * @param context
-	 * @return
-	 * @return List<String>
-	 */
-	public static List<String> getContextColumnNames(String context) {
-		if (Util.isEmpty(context, true)) {
-			return new ArrayList<String>();
-		}
-		String START = "\\@";  // A literal "(" character in regex
-		String END   = "\\@";  // A literal ")" character in regex
-
-		// Captures the word(s) between the above two character(s)
-		String patternValue = START + "(#|$){0,1}(\\w+)" + END;
-
-		Pattern pattern = Pattern.compile(patternValue);
-		Matcher matcher = pattern.matcher(context);
-		Map<String, Boolean> columnNamesMap = new HashMap<String, Boolean>();
-		while(matcher.find()) {
-			columnNamesMap.put(matcher.group().replace("@", "").replace("@", ""), true);
-		}
-		return new ArrayList<String>(columnNamesMap.keySet());
-	}
-
-	/**
-	 * Determinate if columnName is used on context values
-	 * @param columnName
-	 * @param context
-	 * @return boolean
-	 */
-	public static boolean isUseParentColumnOnContext(String columnName, String context) {
-		if (Util.isEmpty(columnName, true)) {
-			return false;
-		}
-		if (Util.isEmpty(context, true)) {
-			return false;
-		}
-
-		// @ColumnName@ , @#ColumnName@ , @$ColumnName@
-		StringBuffer patternValue = new StringBuffer()
-			.append("@")
-			.append("($|#){0,1}")
-			.append(columnName)
-			.append("(@)")
-		;
-
-		Pattern pattern = Pattern.compile(
-			patternValue.toString(),
-			Pattern.CASE_INSENSITIVE | Pattern.DOTALL
-		);
-		Matcher matcher = pattern.matcher(context);
-		boolean isUsedParentColumn = matcher.find();
-
-		// TODO: Delete this condition when fix evaluator (readonlyLogic on Client Info)
-		// TODO: https://github.com/adempiere/adempiere/pull/4124
-		if (!isUsedParentColumn) {
-			// @ColumnName , @#ColumnName , @$ColumnName
-			patternValue.append("{0,1}");
-			Pattern pattern2 = Pattern.compile(
-				patternValue.toString(),
-				Pattern.CASE_INSENSITIVE | Pattern.DOTALL
-			);
-			Matcher matcher2 = pattern2.matcher(context);
-			isUsedParentColumn = matcher2.find();
-		}
-
-		return isUsedParentColumn;
-	}
-
-}
