+++ src/main/java/org/spin/base/util/RecordUtil.java
@@ -51,6 +51,7 @@ import org.spin.backend.grpc.common.Entity;
 import org.spin.backend.grpc.common.ListEntitiesResponse;
 import org.spin.base.db.FromUtil;
 import org.spin.base.db.ParameterUtil;
+import org.spin.dictionary.util.DictionaryUtil;
 import org.spin.service.grpc.util.value.ValueManager;
 
 import com.google.protobuf.Struct;
