+++ src/main/java/org/spin/server/AllInOneServices.java
@@ -30,7 +30,6 @@ import org.spin.grpc.service.BusinessData;
 import org.spin.grpc.service.BusinessPartner;
 import org.spin.grpc.service.CoreFunctionality;
 import org.spin.grpc.service.Dashboarding;
-import org.spin.grpc.service.Dictionary;
 import org.spin.grpc.service.Enrollment;
 import org.spin.grpc.service.FileManagement;
 import org.spin.grpc.service.GeneralLedger;
@@ -57,6 +56,7 @@ import org.spin.grpc.service.UserCustomization;
 import org.spin.grpc.service.UserInterface;
 import org.spin.grpc.service.WebStore;
 import org.spin.grpc.service.Workflow;
+import org.spin.grpc.service.dictionary.Dictionary;
 import org.spin.grpc.service.field.location_address.LocationAddress;
 import org.spin.grpc.service.form.ExpressMovement;
 import org.spin.grpc.service.form.ExpressReceipt;
